pipelines/newsletter/requirements.txt
# Dependencies for Newsletter PrizmAI
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# RSS and web scraping
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Email and HTML
mailchimp3>=3.0.21
premailer>=3.10.0
Jinja2>=3.1.2

# AI and NLP (if using OpenAI)
openai>=1.0.0

# Data processing
pandas>=2.0.0

# Configuration and utilities
pyyaml>=6.0
pathlib2>=2.3.7

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0

# Optional: for advanced text processing
nltk>=3.8
textstat>=0.7.0

# Note: sqlite3 is built-in with Python